+++ tests/integration/ip_messaging/v1/test_credential.py
@@ -39,15 +39,15 @@ class CredentialTestCase(IntegrationTestCase):
                         "sandbox": "False",
                         "date_created": "2015-10-07T17:50:01Z",
                         "date_updated": "2015-10-07T17:50:01Z",
-                        "url": "https://ip-messaging.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                        "url": "https://chat.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ],
                 "meta": {
                     "page": 0,
                     "page_size": 1,
-                    "first_page_url": "https://ip-messaging.twilio.com/v1/Credentials?PageSize=1&Page=0",
+                    "first_page_url": "https://chat.twilio.com/v1/Credentials?PageSize=1&Page=0",
                     "previous_page_url": null,
-                    "url": "https://ip-messaging.twilio.com/v1/Credentials?PageSize=1&Page=0",
+                    "url": "https://chat.twilio.com/v1/Credentials?PageSize=1&Page=0",
                     "next_page_url": null,
                     "key": "credentials"
                 }
@@ -68,9 +68,9 @@ class CredentialTestCase(IntegrationTestCase):
                 "meta": {
                     "page": 0,
                     "page_size": 1,
-                    "first_page_url": "https://ip-messaging.twilio.com/v1/Credentials?PageSize=1&Page=0",
+                    "first_page_url": "https://chat.twilio.com/v1/Credentials?PageSize=1&Page=0",
                     "previous_page_url": null,
-                    "url": "https://ip-messaging.twilio.com/v1/Credentials?PageSize=1&Page=0",
+                    "url": "https://chat.twilio.com/v1/Credentials?PageSize=1&Page=0",
                     "next_page_url": null,
                     "key": "credentials"
                 }
@@ -110,7 +110,7 @@ class CredentialTestCase(IntegrationTestCase):
                 "sandbox": "False",
                 "date_created": "2015-10-07T17:50:01Z",
                 "date_updated": "2015-10-07T17:50:01Z",
-                "url": "https://ip-messaging.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "url": "https://chat.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
         ))
@@ -142,7 +142,7 @@ class CredentialTestCase(IntegrationTestCase):
                 "sandbox": "False",
                 "date_created": "2015-10-07T17:50:01Z",
                 "date_updated": "2015-10-07T17:50:01Z",
-                "url": "https://ip-messaging.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "url": "https://chat.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
         ))
@@ -174,7 +174,7 @@ class CredentialTestCase(IntegrationTestCase):
                 "sandbox": "False",
                 "date_created": "2015-10-07T17:50:01Z",
                 "date_updated": "2015-10-07T17:50:01Z",
-                "url": "https://ip-messaging.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "url": "https://chat.twilio.com/v1/Credentials/CRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
         ))